refactor addbatch to own component
